Move qa_port_array from separate test into qa_Block
